fix: hide "../" when in root dir
Currently file_browser will always insert an "up dir" entry (../),
regardless of where it is in the filesystem

This change suppresses this entry if we are in the root directory (/)

Note I do not have access to Windows. I don't think \ is ever a
valid path in Windows, so don't believe this will have any adverse
affects on that OS, but I haven't been able to check specifically.