feat: change directory based on spcieal input paths
This change adds the following 2 special actions based on the input
path:

- If the user enters `~/` in the TelescopePrompt buffer, the
  file_browser is automatically changed to the user's `${HOME}`
directory

- If the user enters `/` in the TelescopePrompt buffer, the file_browser
  is automatically changed to the root directory (`/`)

The latter action (switching to root) clashes with the existing behaviour
where entering `/` into the TelescopePrompt buffer switches to the directory
browser.

Since this behaviour is also available under `<C-f>` (in insert mode) and
`f` (in normal mode), losing this special action was deemed acceptable.