FIX - Switch between file/folder mode (#43)
Before this commit, switching between file and folder browser mode
makes the finder return to the 'cwd' directory.

This behavior seems a bit odd, with this commit, instead of returning to
'cwd', the finder will remain in the current directory after switching
modes, this way the user can switch from file/folder mode without having
to navigate to the same directory again.

If a user wants to go back to 'cwd' they can always use 'C-W' to do
that.