add feat: option `select_buffer`
If the option is set to `true`, file browser will try to select the
current buffer at start. It has no effect if `opts.path` or `opts.cwd`
is not the parent directory of the current buffer or `opts.depth` is
greater than 1 or the initial mode is not `file`.