fix(action): prevent collapse_dirs from trying to open file (#351)
Previously, opening a directory with one file with `collapse_dirs` set
to true would result in the `open_dir` action trying to refresh the
picker with that file as the new path. This caused `fd` to error out
since a file isn't a valid folder. This was caused by a bad truthy
check.

Corrected this behavior to open folders regularly with `collapse_dirs`
if the selected folder only contains one file.

closes #348