fix(git): show status for git submodules
Need to compute the git toplevel directory upon changing directories
since calling `git status` inside a git submodule will give the status
relative to the toplevel directory of the submodule, not the parent
toplevel directory. This was causing a bug with bad paths being created
for git status changes in submodules.

closes #353