fix(actions.move): `scan_dir` folding moving (#358)
`plenary.scan_dir`, unlike `fd` doesn't append trailing path separators
to directory paths. This was leading to inconsistent behaviors getting
the basename of directories depending on the usage of scan_dir/fd.

This PR refactors the entry maker to standardizes path strings to NOT
end with path separators regardless of type. This is inline with
`fs_realpath` and more compatible with `vim.fs` functions.